+++ src/main/java/com/example/demo/src/support_comment/model/GetSupportCommentRes.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class GetSupportCommentRes {
-    private Long support_comment_idx;
-    private Long support_idx;
-    private Long user_idx;
+    private Integer support_comment_idx;
+    private Integer support_idx;
+    private Integer user_idx;
     private String support_comment_content;
     private String support_comment_updateAt;
 }
