+++ src/main/java/com/example/demo/src/support_comment/model/PostSupportCommentReq.java
@@ -9,8 +9,8 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class PostSupportCommentReq {
-    private Long support_idx;
-    private Long user_idx;
+    private Integer support_idx;
+    private Integer user_idx;
     private String support_comment_content;
 }
